+++ hyp3_gamma/__main__.py
@@ -1,5 +1,5 @@
 """
-rtc_gamma processing for HyP3
+rtc_gamma and insar_gamma processing for HyP3
 """
 import logging
 import os
@@ -114,9 +114,9 @@ def rtc():
 
 def phase_filter_valid_range(x: str) -> float:
     x = float(x)
-    if 0.0 < x <= 1.0:
+    if 0.0 <= x <= 1.0:
         return x
-    raise ValueError(f'{x} not in range (0.0, 1.0]')
+    raise ValueError(f'{x} not in range [0.0, 1.0]')
 
 
 def insar():